make push!

+++ includes/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: gbuczyns <gbuczyns@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 17:44:12 by ssuchane          #+#    #+#             */
-/*   Updated: 2024/11/10 17:53:53 by gbuczyns         ###   ########.fr       */
+/*   Updated: 2024/11/10 18:07:42 by gbuczyns         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,6 +71,6 @@ int		mouse_press(int button, int x, int y, void *param);
 int		mouse_release(int button, int x, int y, void *param);
 int		mouse_move(int x, int y, void *param);
 void	setup_controls(t_display *data);
-int		scale_color(int color, int factor);
+int		scale_color(int color, float factor);
 
 #endif
